Calendar applet: Update each second if global time format is set to include seconds
The code from 044fd02 updated each second if a custom time format was
used only, but in the Date & Time configuration, the default time format
can be modified to include seconds as well. This commit detects this by
checking whether there are at least two colons in the formatted time
string. (It is required to re-check this on each update to notice
changes in the configuration, as the Date & Time configuration does not
notify the applet of the change.)

This is related to PR #4361 and bug #4588.