Systray applet: Fixes for invisible/out-of-place icons
Previously we were resizing icons to force them to redraw. This was messy and it didn't really work.

While testing this, it turned out some of the icons could be fixed by simply hiding/showing them. This worked for some but not for most of them.

For most icons the solution was to delay the insertion in the systray by 1 second.

Here are the special cases which were found while working on this:

- Shutter:
   . Was invisible without disk cache (test with "echo 3 | sudo tee /proc/sys/vm/drop_caches" to empty the cache) --> fixed
   . Looked out of place (small icon on top of big icon) in panel custom size mode --> not resized anymore, looks nice but small now.

- Pidgin:
  . Was invisible without disk cache --> fixed (had to delay by a whole 10 sec.. showing the systray too early made it look invisible)
  . Looked out of place in custom size mode --> fixed

- Dropbox:
  . Didn't scale at all --> Still doesn't.

- Filezilla:
  . Looked out of place in custom size mode --> not resized anymore, always looks small

This might also fix icons such as blueberry or mintupdate.