Preferred application settings: Set mimetypes more exhaustively
If the user selects an app for "Music", he/she expects all the music files to be opened by that app. Same thing for Video, and video files..

Apps don't accurately announce support for all known mimetypes though.. for instance Xplayer announces audio/x-mp3, but but audio/mp3.

This was leading to cases where setting apps would set mimetypes which couldn't be set again for other apps.

For instance, if you set Banshee as the music player, it's assigned audio/mp3... but even if you change your choice to Xplayer, audio/mp3 remains with Banshee.