Show flags based on short name of language in keyboard applet
Keyboard applet showed the flag for the base language even if a foreign
language variant was chosen, e.g. "Russian (Sweden, phonetic)" was shown
with a Swedish flag even when the Russian flag is a more reasonable choice.

Also renamed the Esperanto and Braille icons to match their short names.