xlet settings: Replace old binding function in settings.js with two new ones and allow multiple bindings on the same key
The primary reason for this is that the binding direction is no longer meaningful internally due to the use of javascript properties rather than using a watch function. In addition, there are many instances where it is preferable to bind a key to a different object than the rest, and even to multiple objects.

In its place are two new functions:
* 'bind' works basically the same as 'bindProperty' did, but without the need to pass a binding direction as the first argument.
* 'bindWithObject' works identically to 'bind' but allows the caller to specify the object to which the property will be bound.

Three new functions were also added to unbind keys:
* 'unbind' removes a binding on the bind object passed to the settings object on initialization.
* 'unbindWithObject' removes a binding to a specific object (usually when bound using 'bindWithObject'.
* 'unbindAll' removes all bindings on a single key.

'bindProperty' and 'unbindProperty' are now considered deprecated. They still work as wrappers around 'bind' and 'unbind' respectively, but it is recommended that 'bind', 'bindWithObject', 'unbind', 'unbindWithObject', and 'unbindAll' be used in newer code.

The system applets and desklets were updated to reflect these changes.