cs_backgrounds.py: Use built in mimetypes module
In my testing, using mimetypes instead of subprocess seems to give a
huge speed boost, especially for the cached file loading.

This commit also removes the duplicate mimetype check in
update_icon_view as it causes UI hangs while checking a large amount
of files. We check it again during the actual background loading in
get_pix, which should be good enough.