cs_backgrounds.py: Refactor disk caching
This changes the cache file format from a normal png file to
a pickled object which contains a png file and the original
image dimensions.

There are two benefits to this method:

1) Allows us to stop loading the original file every time
just to get the dimensions for display purposes.
2) Svg files can now be loaded properly after the initial run.

Previously, svg files would not be displayed after the initial
run as we were using PIL to get the dimensions, and it does not
handle svg files.

Finally, the use of temp files to convert between pixbuf and PIL
images has been eliminated as gdk-pixbuf now provides a reliable
method to load a pixbuf from bytes in memory.