window-list: Rework for vertical panels
Currently the window list items are extremely small and centered within vertical panels which isn't very nice visually and not very useful. Rework the layout so the they fill the full horizontal width. Also scale the icons based on panel size so they are essentially square instead of tiny rectangles.