cs_themes: Don't show duplicate themes
Currently when we have a theme of the same name in both the /usr and /home
directory both themes are shown in the theme chooser buttons. When choosing either
of the themes in the settings the one in /home will be given precedence. When
both exist, only show the one from the users home directory since that is the
one that will be applied regardless of which the user chooses.

Fixes https://github.com/linuxmint/Cinnamon/issues/5778