[menu-applet] decodeUriComponent in placesManager.js and docInfo.js
The app description in the menu-applet doesn't handle some characters (e.g. german umlaute ö ü ä) in URIs properly.
To fix this, two variables, idDecode and uriDecode, were added to the above mentioned files.
This variables are decoded URIs. They can be used, when an URI has to be shown properly in a string.

Also the slice of the URIs wasn't handled properly in the menu-applet.
It was assumed, that a bookmark in the Places category of the menu always starts with bookmark:file://
That's not the case, e.g. when you mount an USB-Stick, the USB-Stick is shown in the menu as mount:file://
To solve this, the description of the app now cuts everything out, what appears before the first colon : in the URI string of a bookmark.
If it's a file, then also cut file:// out.
(That's also changed for the Recent files category.)

Fixes https://github.com/linuxmint/Cinnamon/issues/5595, fixes https://github.com/linuxmint/Cinnamon/issues/3620
and fixes https://github.com/linuxmint/Cinnamon/issues/3250