[Menu applet] Added keyboard navigation for context menus.
The latest changes to the menu applet added the possibility to open the context
menu for menu items with the keyboard, but the context wasn't navigable. This
pull request adds the ability to navigate the context menus with the keyboard.

Usage:
- The Arrow Up/Down and PageUp/PageDown keys will navigate through the menu
items of an open context menu.
- If the Arrow Left/Right or Tab keys are pressed, it will continue navigation
outside the applications box.
- If the Escape key is pressed while a context menu is open, first it will close
the opened context menu and if it's pressed again it will close the applet menu.