[menu-applet] hide recent files, which are no longer available
Recent files, which are no longer available, will be hidden in the recent files category after next call of refreshRecent.
If the file is available again, because it was on a CD or USB, it will appear again.
If a file is no longer available, but the recent files are not refreshed yet, the description of the file shows a warning "This file is no longer available".