[Xlet settings] "entry" and "iconchooser" elements modifications.
By default, "entry" and "iconchooser" editable fields are too small on xlet's
setting windows. In addition, they don't expand their width when the window
is resized. This makes those elements *uncomfortable to work with*.

This commit adds to the "entry" and "iconchooser" elements a new boolean option
(optional) called "expand-width". When set to **true**, the editable fields for
"entry" and "iconchooser" elements will occupy all available space in the row
and at the same time will expand if the window is resized.