[Tooltips] Fixed positioning of tooltips.
Currently, tooltips created with the **new Tooltips.Tooltip()** constructor
are aligned to the top-left corner of the mouse pointer. This leads to having
tooltips overlapped by the mouse pointer. And the bigger the mouse pointer, the
more of the tooltip is covered by it.

This commit aligns the tooltip to the bottom-right corner of the mouse pointer
(approximately), reducing the possibility of the mouse pointer to overlap
the tooltip.

Corrected some minor indentation/white space inconsistencies.

This commit fixes linuxmint/Cinnamon#5851