popupMenu: Insert menus that open from the panel below the panels
Do this by first testing if the menu is originating from the panel or a source
actor that is located on the panel. If it is, grab all the panels located on the
monitor and insert the menu just below the lowest panel. This prevents weird
visual artifacts where the menu can slightly overlap the panel in many themes.