[cinnamon-json-makepot] Added support for "title" key
With the redesign of the xlets settings system, there is a new UI element called
"page". These elements have a key called "title" that needs to be translated.

This commit adds the ability to the `cinnamon-json-makepot` command to extract
from the settings-schema.json file those keys when said command is used to
create a .pot file.