[applet.js] Re added ability to display a separator on applets context menus.
In Cinnamon version 3.0.7, when new menu items are added to the context menu
of an applet/desklet, a separator automatically appears, dividing the default
menu items (**About**, **Configure...** and **Remove...**) from the just added
menu items. In Cinnamon version 3.2.x, the same behavior was observed until
the https://git.io/v6rI7 commit removed it.

This commit brings back that behavior to the context menu for applets.