Add cinnamon-stap-monitor utility.
Needs quite a bit more work, but is useful as is for now.

Usage: (run from the folder these files are in) -
    cinnamon-stap-monitor <program to monitor>

like:

cinnamon-stap-monitor nemo

or

cinnamon-stap-monitor "cinnamon --replace"

Requires systemtap to be installed and functional (tested on
fedora 25), as well as glib compiled with stap/dtrace support -
once I figure out how to attain this in Mint/Ubuntu I will write
a tutorial - it's super broken right now though.