JS API misc/util.js: Add spawnCommandLineAsync
This is a CLI utility method that invokes a callback after the process
has exited cleanly, or an error callback if it does not. Neither
callback will be called if the functions are not passed to the method.
Having callbacks allows more precise handling of the asynchronous
behavior involved with calling processes from JS.

This uses GLib.child_watch_add to return the callbacks after the process
exits.
https://people.gnome.org/~gcampagna/docs/GLib-2.0/GLib.child_watch_add.html