cinnamon-settings: Fix some sliders being drawn backwards
Add a flipped argument to the cinnamon settings range widget. Gtk.Scale is a
little limited in what it allows. There is no way built in to map a range from
a big->small value without drawing the fill value in reverse. Add a bit of a
hack to allow working around this limitation.