hotcorners: remove the icon option
This is a left over legacy from gnome shell and simply doesn't work well the
various panel layouts we allow. It just appears broken/unfinished.

Remove the ability to use them and add a new gsettings key to manage our
settings for them. This should ensure nothing gets broken when Cinnamon is
upgraded.