panel.js: refactor _moveResizePanel
We now check for changes instead of unconditionally moving and resizing
panels. When there are changes we now position the panel while taking the
visibility state into account. This fixes almost all cases of hidden panel
flickering.

The clip region is now updated during _moveResizePanel to ensure that we only
have a shadow on one side of the panel. This prevents shadow overlap in
certain cases, but not all, and also prevents shadows from showing up on
adjacent monitors.

Another new behavior, is that we now remove all tweens from the panel if it
has changed. This prevents tweens from running during repositioning which
caused panels to sometimes disappear if they were moved during an animation.