melange/cinnamon-looking-glass.py: convert to normal toplevel window
This converts the Melange window to a normal, non-docked, top level window
so it can placed wherever you want. In addition to making the window useful
in more cases, this fixes an existing bug where the window is misplaced and
missized when vertical panels are present.

I could not find a way to reliably detect if the window has focus when
the looking glass shortcut is pressed. Because of this, we are unable to
determine if the window needs to show, hide, or present. As a workaround
the window still defaults to always-on-top. You can toggle always-on-top
via the window menu.