main.js: improve logging of GLib.Errors and objects (#6486)
This ensures that GLib.Errors are handled the same as regular Errors
when passed to logging functions, and also improves the usability of
logging functions when called with non-strings.

Previously, calling these logging functions with a non-string would
print the string representation of the object to the file, but would
break logging inside Melange. The stored log objects inside Main must
be strings in order to be passed over dbus successfully to Melange,
or any other dbus caller.

GLib.Error messages were not previously printed, just the string
representation of the object. In addition there  was no stack trace
associated with that error log.