[cinnamon-json-makepot] Added support to extract "title" and "units" keys on list elements
With the introduction of list widgets to the xlets settings system (#6303), there are the new *title* and *units* keys that needs to be extracted to be translated.

This commit adds the ability to the `cinnamon-json-makepot` command to extract from the settings-schema.json file those keys when said command is used to create a .pot file.