workspace.js: initalize WindowClone._zoomStep to 0
We check if this._zoomStep > 0 on down scroll to determine if we need to zoom
out or end the zoom operation, but this is never initialized. It is also set
to undefined after scrolling ends. This causes a harmless undefined reference
warning, but initializing to 0, and checking that instead of undefined to
determine state avoids the warning.