cinnamon-json-makepot: Add keyword option, change language used by xgettext to JavaScript
This changes two things.

- It adds a keyword option that can be passed to xgettext. For instance,
if you don't assign gettext to an underscore variable ("_"), setting
this option with the alternative variable will allow the JS strings to
be included.

- The script breaks on my applet's files because of regex found in the
Lodash library. This is fixed by changing the language to "JavaScript".
If there's side effects of this I'm not aware of, I can make this an
option too, but it seems like a safe choice.