popupMenu.js: Stop relying on an idle callback to toggle menus.
This could occasionally cause menus to become unresponsive, or very slow
to respond, since an idle callback won't be called until no higher priority
events are pending.

In addition to removing the idle call, fix a few places where we were (now) doing
menu calculations *after* the animation had started, rather than before.

There remains some small positioning issue, where depending on what is populating
a menu and what's visible or not in it can cause a slight difference in the menu's
width (or height for vertical) from the pre-visible calculated natural size, to the
actual mapped size.  The previous behavior was, during our open animation, we blocked
the normal "allocate" signal handler (by monitoring the "animating" property), ran
_calculatePosition() once, set the menu's position, and ran the animation.  Once the
animation completed, the allocate handler was unblocked, and _calculatePosition() was
run again, but based off of (for whatever reason) a slightly different width/height,
causing a small but perceptible shift (half of the width difference, 2 pixels in my
problem case).

New behavior is to, during the onUpdate callback, check against the original width/height
of the actor, and if its changed, shift the actor to the correct position - this is more
or less imperceptable because it will happen on the first tick of the animation.

- Also removed the opacity portion of the menu animations.