On-screen keyboard: Remove animations
For three reasons..

First, this isn't just something we launch, it's also something which reacts to
AT-SPI event, i.e. the fact that a widget gets focus can result in showing
the on-screen keyboard. When that happens, things should move fast and look
immediate.

Second, this affects struts and thus the position and maximization of windows...
here again, we want changes to be immediate. Failing that, what we see is the empty
part of the desktop which is freed (beyond struts) and behind the keyboard.. as the
keyboard isn't fully opaque yet and about to be shown.

Third, there seems to be some race condition affecting struts. In some cases hiding
the keyboard doesn't free space and the keyboard box still claims it.. making it impossible
to fully maximize windows.