Performance: Avoid caching `this` context in SignalManager instances (#6918)
This passes an null so SignalManager doesn't cache the context,
which it uses for binding callbacks. This can be replaced by
SignalManager.connect's fourth parameter, bind.

Since the manager is a property of its callee, re-referencing `this`
inside each SignalManager instance will create a circular reference, and
while they are not inherently bad, it forces the engine to check more 
objects. Spidermonkey in particular counts circular references as
part of an object's reference count, which makes it more likely an
object will exist after its needed, and it will continue to consume
memory.