l10n: Update POT file and fix some msgids
In many Latin languages, it is not possible to translate an adjective or
sometimes a past-participle witout knowing the noun it relates to.

For instance: "Available %s" cannot be translated. "Available" differs
depending on the gender and plurality of %s.

For this reason we cannot use generic sentences such as "Available %s".
Instead we need to be more precise and use "Available applets",
"Available themes" etc..

Same thing with "%s could not be removed", or "Error while removing %s, remove it manually".
Im the former example, "removed" cannot be translated because we don't know its gender.
In the latter example "it" cannot be translated, for the same reason.