Sound settings: Assign the selected device to the profile selector.
When the user selects a device, we ask Cvc to switch the output/input
to that device. That can either fail, in which case nothing happens,
or succeed, in which case we get a signal from Cvc telling us the
active device changed.

When we get that signal, we update our profile selector.

The problem is that sometimes, setting the active device will fail
because of the profile we selected, and because it fails, it never becomes
the active device and so we cannot set the profile for that device again.

We get into a situation where, after selecting the wrong profile for a
particular device, we can no longer change the profile for that device
and thus setting it as the active device anymore.

This commit simply sets the profile selector to the selected device.
That way we're able to access its profiles, without relying on whether
or not it succeeded in becoming the active device.